Fix CI: too-many-lines lint, and address Codex review findings
- Fix clippy::too_many_lines on the new SSE migration-resume test
  (132/100 lines), matching the #[allow] already present on the
  handler it exercises.

- P1 (Codex): send_rows's caller-side `rows.last()` read for
  last_seen_event_id ignored should_break, reporting a row the
  channel never sent when try_send failed partway through a batch.
  A reconnecting client's drop_after_event_id then pointed past
  events it never received. Track last_seen_event_id inside
  send_rows itself, alongside last_seen_id, updated only after a
  row is actually sent; return both from the closure instead of
  reconstructing from the input slice at the call site.

- P2 (Codex): the public SSE cursor contract still described
  Last-Event-ID as harvest_events.id/`id > n` in docs/api-contract.json
  (and its generated OpenAPI artifacts), docs/management-api.md, and
  the CLI's `harvest events tail --last-event-id` help/type. Updated
  all four to describe the event_id cursor, regenerated the OpenAPI
  JSON files via scripts/regenerate-openapi.sh, and narrowed the CLI
  flag from i64 to i32 to match the server's actual accepted range.

- Also corrects several comment-hygiene sentence-length violations
  this session's earlier local runs missed: a stale git ref made an
  earlier local check compare against the wrong merge base (166
  files instead of the true 7), so several long sentences were never
  actually validated. Fixed now against the correct base.

+++ autumn-harvest-cli/src/lib.rs
@@ -3417,10 +3417,13 @@ enum EventsCommand {
     Tail {
         /// Workflow execution ID to watch.
         execution_id: String,
-        /// Resume from this event row ID (Last-Event-ID header).
-        /// Events with id > this value are replayed before entering live-tail mode.
+        /// Resume from this event_id (issue #1405), sent as the
+        /// Last-Event-ID header. NOT the shard-local harvest_events.id --
+        /// this is the per-execution sequence number the server's `id:`
+        /// SSE field carries. Events after it are replayed before entering
+        /// live-tail mode.
         #[arg(long)]
-        last_event_id: Option<i64>,
+        last_event_id: Option<i32>,
     },
 }
 
@@ -7685,7 +7688,7 @@ pub async fn execute(cli: &Cli) -> Result<Value, CliError> {
 async fn run_events_tail(
     cli: &Cli,
     execution_id: &str,
-    last_event_id: Option<i64>,
+    last_event_id: Option<i32>,
 ) -> Result<(), CliError> {
     let path = format!("/executions/{}", path_segment(execution_id));
     let url = format!(
